Replace AsyncParallelPartitioningCollector with AsyncParallelCollector
